// frontend/src/app/models/user.ts
export class User {
  id?: number;
  first_name?: String;
  last_name?: String;
  email?: String;
  password?: String;
  is_active?: boolean;
  is_superuser?: boolean;
  is_employed?: boolean;
  date_of_employment?: Date;
  date_of_fired?: Date;
  company_id?: number;
  employment_id?: number;
  position_id?: number;
}